requirements.txt
python-telegram-bot==21.10
python-dotenv>=0.21.0

#jupyter
#pandas

#scikit-learn

#joblib

#seaborn

#pytest
#pytest-cov
#black
#pylint
#mypy

#fastapi

#uvicorn[standard]

#requests

#starlette-exporter

#mlflow
#boto3